+++ Utils.cpp
@@ -1,0 +1,19 @@
+#include "Utils.h"
+#include <Windows.h>
+
+std::string Utils::WStringToUtf8(const std::wstring& wstr)
+{
+    if (wstr.empty()) return {};
+
+    int size_needed = WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), -1, nullptr, 0, nullptr, nullptr);
+    if (size_needed <= 0)
+    {
+        return {};
+    }
+
+    std::string utf8str(size_needed, 0);
+    WideCharToMultiByte(CP_UTF8, 0, wstr.c_str(), -1, &utf8str[0], size_needed, nullptr, nullptr);
+
+    utf8str.pop_back();
+    return utf8str;
+}
